refactored lots of Rcpp functions
